chore(Canvas): Split CanvasService in Controller and Flow services
